requirements.txt
-i https://pypi.org/simple
typer
python-dotenv
typing-extensions; python_version >= '3.7'
delta-spark==2.4.0
pyspark==3.4.1
